+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -15,7 +15,7 @@
 #import "SentryAppStartTrackingIntegration.h"
 #import "SentryAppState.h"
 #import "SentryAppStateManager.h"
-#import "SentryAttachment.h"
+#import "SentryAttachment+Private.h"
 #import "SentryAutoBreadcrumbTrackingIntegration+Test.h"
 #import "SentryAutoBreadcrumbTrackingIntegration.h"
 #import "SentryAutoSessionTrackingIntegration.h"
@@ -176,6 +176,8 @@
 #import "UIViewController+Sentry.h"
 #import "URLSessionTaskMock.h"
 @import SentryPrivate;
+#import "SentryEnvelopeAttachmentHeader.h"
+#import "TestSentryViewHierarchy.h"
 
 #if SENTRY_HAS_UIKIT
 #    import "MockUIScene.h"